requirements.txt
aiogram==3.20.0.post0
aiohttp==3.11.18
SQLAlchemy==2.0.41        
asyncpg==0.30.0
aiosqlite==0.18.0       
aiocache==0.12.2
redis==6.1.0
aioredis==2.0.1
pydantic>=2.7,<2.12        
pydantic-settings>=2.9
python-dotenv==1.1.0
structlog==25.3.0
orjson==3.10.18
#uvloop==0.21.0
celery==5.5.2
backoff==2.2.1           
httpx==0.24.1
loguru==0.7.0
PyYAML==6.0               
APScheduler==3.10.4
aiolimiter
